Split planner DSL by helper role
- Move expression DSL helpers into focused child modules behind the existing facade.
- Move FunctionDsl params, steps, and return-body helpers into child modules.
- Keep existing planner::dsl import surface while colocating DSL unit tests by module.
- Document the wildcard import rule for facade-only re-export exceptions.
