+++ src/planner/dsl/function/params.rs
@@ -1,0 +1,124 @@
+use super::FunctionDsl;
+use crate::plan::{
+    BoolFunctionLocalId, BoolLocalId, FunctionFunctionLocalId, FunctionType, IntFunctionLocalId,
+    IntLocalId, NilFunctionLocalId, NilLocalId, Param, ParamLocal, StringFunctionLocalId,
+    StringLocalId, ValueType,
+};
+use ecow::EcoString;
+
+impl FunctionDsl {
+    pub(crate) fn param_int(mut self, local: usize, name: impl Into<EcoString>) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::int(IntLocalId(local)),
+            name.into(),
+        ));
+        self
+    }
+
+    pub(crate) fn discard_int_param(mut self, local: usize) -> Self {
+        self.params
+            .push(Param::discard(ParamLocal::int(IntLocalId(local))));
+        self
+    }
+
+    pub(crate) fn param_string(mut self, local: usize, name: impl Into<EcoString>) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::string(StringLocalId(local)),
+            name.into(),
+        ));
+        self
+    }
+
+    pub(crate) fn param_bool(mut self, local: usize, name: impl Into<EcoString>) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::bool(BoolLocalId(local)),
+            name.into(),
+        ));
+        self
+    }
+
+    pub(crate) fn param_nil(mut self, local: usize, name: impl Into<EcoString>) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::nil(NilLocalId(local)),
+            name.into(),
+        ));
+        self
+    }
+
+    pub(crate) fn param_int_function(
+        mut self,
+        local: usize,
+        name: impl Into<EcoString>,
+        arguments: impl IntoIterator<Item = ValueType>,
+    ) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::int_function(
+                IntFunctionLocalId(local),
+                FunctionType::new(arguments.into_iter().collect(), ValueType::Int),
+            ),
+            name.into(),
+        ));
+        self
+    }
+
+    pub(crate) fn param_string_function(
+        mut self,
+        local: usize,
+        name: impl Into<EcoString>,
+        arguments: impl IntoIterator<Item = ValueType>,
+    ) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::string_function(
+                StringFunctionLocalId(local),
+                FunctionType::new(arguments.into_iter().collect(), ValueType::String),
+            ),
+            name.into(),
+        ));
+        self
+    }
+
+    pub(crate) fn param_bool_function(
+        mut self,
+        local: usize,
+        name: impl Into<EcoString>,
+        arguments: impl IntoIterator<Item = ValueType>,
+    ) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::bool_function(
+                BoolFunctionLocalId(local),
+                FunctionType::new(arguments.into_iter().collect(), ValueType::Bool),
+            ),
+            name.into(),
+        ));
+        self
+    }
+
+    pub(crate) fn param_nil_function(
+        mut self,
+        local: usize,
+        name: impl Into<EcoString>,
+        arguments: impl IntoIterator<Item = ValueType>,
+    ) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::nil_function(
+                NilFunctionLocalId(local),
+                FunctionType::new(arguments.into_iter().collect(), ValueType::Nil),
+            ),
+            name.into(),
+        ));
+        self
+    }
+
+    pub(crate) fn param_function_function(
+        mut self,
+        local: usize,
+        name: impl Into<EcoString>,
+        type_: FunctionType,
+    ) -> Self {
+        self.params.push(Param::named(
+            ParamLocal::function_function(FunctionFunctionLocalId(local), type_),
+            name.into(),
+        ));
+        self
+    }
+}
